requirements.txt
confluent-kafka
python-dotenv
certifi==2025.4.26
charset-normalizer==3.4.2
# contourpy==1.3.2
contourpy
cycler==0.12.1
fonttools==4.58.0
idna==3.10
kafka-python==2.2.10
# kiwisolver==1.4.8
kiwisolver
matplotlib
numpy==2.2.6
packaging==25.0
pandas==2.2.3
pillow==11.2.1
py4j==0.10.9.9
pyparsing==3.2.3
pyspark==4.0.0
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
six==1.17.0
tzdata==2025.2
urllib3==2.4.0
